basebackup: backup user tablespaces in `local-tar` basebackup mode
The generated files are in a custom, new, format which includes a top-level
JSON metadata file containing tablespace mapping, /pgdata/ for files in
$PGDATA and /tablespaces/name/ directories for all tablespaces.

Restore extracts the default tablespace files to $PGDATA and other
tablespaces to the same directories where they were in the host where the
backup was taken.  These directories can be overridden using the
`--tablespace-dir TSNAME=DIRECTORY` flag for restore.  The directories must
exis and be empty.

+++ test/conftest.py
@@ -54,6 +54,8 @@ class TestPG:
         time.sleep(1.0)  # let pg start
 
     def kill(self, force=True):
+        if self.pg is None:
+            return
         os.kill(self.pg.pid, signal.SIGKILL if force else signal.SIGTERM)
         timeout = time.time() + 10
         while (self.pg.poll() is None) and (time.time() < timeout):
